sacrifice a bit of diff granularity for a huge improvement in performance
Diffing by character, as is the default in ReactDiffViewer, can cause a
delay of several seconds before displaying just one diff. The "Apply
All" call was showing each new diff before changing it, but ironically
the diff viewing took orders of magnitude more time than the actual
application of the change. I assumed that when you're applying to all
you're relatively confident in the change and don't need to view them go
by. So I removed the behavior of showing the diffs as processing
happened. The character diffing was swapped out for line-based diffing
to speed up the case of going one-by-one through the changes and
reviewing them before applying them. More performant diffing approaches
could be considered in the future so we can keep the character-based
diff.